+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 ## 0.58.0 (Unreleased)
-ENHANCEMENTS:
-* make `standard-v2` a default platform for `yandex_compute_instance` and `yandex_compute_instance_group` resources.  
+FEATURES:
+* **New Data Source:** `yandex_alb_target_group`
+* **New Resource:** `yandex_alb_target_group`
 
+ENHANCEMENTS:
+* make `standard-v2` a default platform for `yandex_compute_instance` and `yandex_compute_instance_group` resources. 
+  
 ## 0.57.0 (April 29, 2021)
 FEATURES:
 * support k8s node group placement groups both in resource and data source.